[DSCP-60] 'Serialise' instances for Educator datatypes
