+++ disciplina.cabal
@@ -51,8 +51,9 @@ library
                      , Disciplina.DB.DSL.Interpret.SimpleTxDB
 
                      -- Core datatypes and functions
-                     , Disciplina.Core.Types
                      , Disciplina.Core.ATG
+                     , Disciplina.Core.Serialise
+                     , Disciplina.Core.Types
 
                      -- Witness node model
                      , Disciplina.WorldState.Internal
@@ -60,7 +61,9 @@ library
 
                      -- Educator node model
                      , Disciplina.Educator.Block
+                     , Disciplina.Educator.Configuration
                      , Disciplina.Educator.Txs
+                     , Disciplina.Educator.Serialise
                      , Disciplina.Educator.State
 
                      , Disciplina.Crypto.ByteArray
@@ -135,6 +138,7 @@ library
                        ViewPatterns
 
   ghc-options:         -Wall
+                       -Werror
                        -fno-warn-orphans
 
 executable disciplina-witness
