+++ src/Disciplina/Educator/State.hs
@@ -1,1 +1,7 @@
-module Disciplina.Educator.State () where
+module Disciplina.Educator.State
+       (
+       ) where
+
+-- import Universum
+
+-- import Snowdrop.Model.Expander (Expander (..), expandRawTxs)
